Do not allow cvs repo to be the current directory
In a simple uplift case of a couple of RCS files, a fairly natural
thing to do is try to import the ,v files from the current directory.
Allowing this leads to a recursion loop, with .-proxy subdirectories
being created under ./module, ad infinitum. Add a check to prevent this.